Colour class operators, some shadow work, TODOs, dos2unix
I've added most of the Colour class operators, but I *really* need to
write some tests for these. Not sure the best way to do this yet.

I added a bit more detail to the TODO list, it would be nice to give
these entries some kind of priority…

Pressing issues are still the shadows, esp. in reference to soft shadows
and dealing with opacity in materials. This is probably the next step.

I dos2unixed a few files too.