Added beginnings of testing framework, worked on Colour, fixed bug
Added Colour(std::vector<Colour>& colours), added something to test this
constructor, and also replaced some AA code to use this constructor, which
also fixed a bug.

Started working on a basic testing framework, but this needs to be
rethough because at present it requires compiling everything twice. Very
stupid. Will work on this.