Reimplemented soft shadows. Need fast jitter...
The random function was 50% of the reason it was slow. I need to find a
faster way to generate random numbers. A precalculated list might work
well. Perhaps rand() is faster than fancy C++11 random? Needs looking
into.

12 million rays per second on my Macbook Pro (late 2013) i7 seems to be
about average. I should be careful of sudden drops in the future.